feat: EasterEgg 컴포넌트화
