+++ src/components/user/EasterEgg.tsx
@@ -8,9 +8,7 @@ import IconButton from '@mui/material/IconButton';
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 import Modal from '@mui/material/Modal';
-import { Container, padding } from '@mui/system';
-import { createTheme, styled } from '@mui/material/styles';
-import { red, green, blue } from '@mui/material/colors';
+import { EasterImgLayout, ModalLayout, theme } from '../utils/muiResponse';
 
 const styleModal = {
   position: 'absolute',
@@ -35,77 +33,9 @@ const styleRow = {
   justifyContent: 'center' /* 중앙으로 이동 */,
 };
 
-const styleImg = {
-  width: '46vw' /* 16x8.5= 136px(default값 16) */,
-  height: '64vh' /* 16x19.5= 312px(default값 16) */,
-  alignSelf: 'center',
-  objectFit: 'contain' /* 비율유지한채, 이미지 잘리지 않게, 남는공간 비워둠 */,
-  content: {
-    laptop: `url(${Simpson})`, //img src from 0 up to mobile
-    mobile: `url(${Simpson})`, //img src from mobile up to tablet
-    tablet: `url(${Simpson2})`, //img src from tablet up to desktop
-    desktop: `url(${Simpson3})`, //img src from desktop up to ~
-  },
-  alt: 'Loading...',
-};
-
 const styleIcon = { position: 'absolute', right: 1, top: 1 };
 
-const Box2 = styled('div')(({ theme }) => ({
-  height: '100vh',
-  padding: theme.spacing(1),
-  [theme.breakpoints.up('desktop')]: {
-    height: '50rem',
-    width: '64rem',
-  },
-  // [theme.breakpoints.up('tablet')]: {},
-  // [theme.breakpoints.up('desktop')]: {},
-}));
-
-const Box3 = styled('div')(({ theme }) => ({
-  height: '100vh',
-  padding: theme.spacing(1),
-  // [theme.breakpoints.between('tablet', 'desktop')]: {
-  //   height: '70vh',
-  //   width: '55vw',
-  // },     /* 필요시 사용, 크기 조정*/
-
-  [theme.breakpoints.up('desktop')]: {
-    height: '420px',
-    width: '400px',
-  },
-
-  // [theme.breakpoints.up('desktop')]: {},
-}));
-
-const theme = createTheme({
-  /* custom MediaQuery */
-  breakpoints: {
-    values: {
-      mobile: 480,
-      tablet: 768,
-      laptop: 0,
-      desktop: 1024,
-    },
-  },
-});
-
-declare module '@mui/material/styles' {
-  /* 타입 스크립트때문에 사용 */
-  interface BreakpointOverrides {
-    xs: false; // removes the `xs` breakpoint
-    sm: false;
-    md: false;
-    lg: false;
-    xl: false;
-    mobile: true; // adds the `mobile` breakpoint
-    tablet: true;
-    laptop: true;
-    desktop: true;
-  }
-}
-
-export function EasterEgg() {
+export function EasterEgg({ img1, img2, img3 }: any) {
   const [open, setOpen] = React.useState(true);
   // const handleOpen = () => setOpen(true);
   const handleClose = () => setOpen(false);
@@ -113,6 +43,22 @@ export function EasterEgg() {
   useEffect(() => {
     console.log(open); /* 모달창 상태 확인 */
   }, [open]);
+
+  const styleImg = {
+    /* props 사용하기 위해 함수형 컴포넌트 안에 배치 */
+    width: '46vw' /* 16x8.5= 136px(default값 16) */,
+    height: '64vh' /* 16x19.5= 312px(default값 16) */,
+    alignSelf: 'center',
+    objectFit:
+      'contain' /* 비율유지한채, 이미지 잘리지 않게, 남는공간 비워둠 */,
+    content: {
+      laptop: `url(${img1})`, //img src from 0 up to mobile
+      mobile: `url(${img1})`, //img src from mobile up to tablet
+      tablet: `url(${img2})`, //img src from tablet up to desktop
+      desktop: `url(${img3})`, //img src from desktop up to ~
+    },
+    alt: 'Loading...',
+  };
 
   return (
     // <Container fixed>
@@ -122,7 +68,7 @@ export function EasterEgg() {
       aria-labelledby="modal-title" /* 모달 제목 */
       aria-describedby="modal-description" /* 모달 설명 */
     >
-      <Box2 sx={styleModal} theme={theme}>
+      <ModalLayout sx={styleModal} theme={theme}>
         <Box sx={styleRow}>
           <Typography
             id="modal-title"
@@ -141,10 +87,10 @@ export function EasterEgg() {
           </IconButton>
         </Box>
         {/* 자체 정렬 */}
-        <Box3 sx={styleImg} theme={theme}>
+        <EasterImgLayout sx={styleImg} theme={theme}>
           <Box component="img" id="modal-description" sx={styleImg}></Box>
-        </Box3>
-      </Box2>
+        </EasterImgLayout>
+      </ModalLayout>
     </Modal>
     //</Container>
   );
