+++ src/vite-env.d.ts
@@ -4,3 +4,18 @@
 interface Window {
   Kakao: any;
 }
+
+declare module '@mui/material/styles' {
+  /* 타입 스크립트때문에 사용 */
+  interface BreakpointOverrides {
+    xs: false; // removes the `xs` breakpoint
+    sm: false;
+    md: false;
+    lg: false;
+    xl: false;
+    mobile: true; // adds the `mobile` breakpoint
+    tablet: true;
+    laptop: true;
+    desktop: true;
+  }
+}
